+++ src/fides/api/main.py
@@ -21,7 +21,15 @@ from uvicorn import Config, Server
 from fides.api.ctl import view
 from fides.api.ctl.database.database import configure_db
 from fides.api.ctl.database.seed import create_or_update_parent_user
-from fides.api.ctl.routes import admin, crud, datamap, generate, health, validate
+from fides.api.ctl.routes import (
+    admin,
+    crud,
+    datamap,
+    generate,
+    health,
+    system,
+    validate,
+)
 from fides.api.ctl.routes.util import API_PREFIX
 from fides.api.ctl.ui import (
     get_admin_index_as_response,
@@ -79,6 +87,7 @@ ROUTERS = crud.routers + [  # type: ignore[attr-defined]
     health.router,
     validate.router,
     view.router,
+    system.router,
 ]